+++ kiltergpt/models/gpt.py
@@ -372,6 +372,47 @@ class GPTModel(L.LightningModule):
         generated = self.generate(x, angle, grade, temperature=temperature, p=p)
         return self.tokenizer.decode(generated, clean=True)
 
+    def generate_generation_fig(
+        self,
+        prompt: str,
+        angle: int,
+        grade: str,
+        temperature: float = 0.7,
+        p: float = 0.8,
+        name: str | None = None,
+        colorscheme: str = "RdPu",
+        cutoff: float = 0.1,
+    ):
+        import imageio
+
+        plotter = Plotter(colorscheme, verbose_cutoff=cutoff)
+        frames = self.generate_from_string(prompt, angle, grade, temperature, p)
+        probs = self._gen_logits
+        probs = probs[0::2]
+        split_frames = [f"p{i}" for i in frames.split("p")[1:]]
+        split_frames = ["".join(split_frames[:i]) for i in range(len(split_frames))]
+        split_frames.append(frames)
+        animation = []
+        for frame, prob in zip(split_frames, probs, strict=True):
+            dict_prob = {}
+            for i in torch.nonzero(prob):
+                dict_prob[self.tokenizer.decode_map[i.item()]] = prob[i].item()
+            animation.append((frame, dict_prob))
+        images = []
+        for i in range(len(animation)):
+            k, v = animation[i]
+            # plot the current frame + options
+            images.append(plotter.plot_climb(k, probs=v))
+            if i < len(animation) - 1:
+                # plot the made choice
+                next_k, _ = animation[i + 1]
+                images.append(plotter.plot_climb(next_k, probs=v))
+        images.append(plotter.plot_climb(frames))
+        if name is None:
+            name = f"{prompt}_{angle}_{grade}_{temperature}_{p}.gif"
+        imageio.mimsave(name, images, fps=1)
+        return frames
+
     @staticmethod
     def load_from_wandb(checkpoint_path: str = "ilsenatorov/model-registry/kiltergpt:best") -> "GPTModel":
         """Use self.load_from_checkpoint to download model weights from wandb. Looks for models in ilsenatorov/kilter-gpt"""